fix(renderer): kitty id 매핑 키를 surface 포인터에서 비재사용 Surface.id 로 (K2e 적대적 1회차)
매핑 키로 surface 주소를 썼다. surface 가 닫히고 두 프레임 사이에 같은 주소로 새
surface 가 할당되면 새 surface 가 옛 매핑(같은 전역 id)을 그대로 이어받는다. 그때
새 이미지가 `id 1·generation 1` 처럼 옛 것과 같은 generation 이면(generation 은
코어마다 1 부터라 가장 흔한 조합) 업로드 기록이 「이미 올렸다」고 답해 옛 surface
의 그림이 보인다. 모듈 주석은 「같은 주소가 재사용돼도 새 매핑으로 시작한다」고
적었는데 **틀린 주장**이었다.

`Surface.id` 는 `SurfaceIdAllocator` 가 단조 증가로 발급하고 재사용하지 않는다
(그 불변식은 `monotonic` 판정자가 이미 고정한다). 키 타입을 `u64` 로 바꿔 포인터가
다시 들어올 수 없게 했다. 주소 재사용은 할당자에 달린 비결정적 현상이라 옛 코드에서도
우연히 초록일 판정자를 세우는 대신 타입과 불변식으로 막는다.

+++ src/renderer/kitty_image_ids.zig
@@ -17,9 +17,12 @@
 const std = @import("std");
 
 pub const Key = struct {
-    /// surface 를 구별하는 값(주소 등). 같은 프레임 안에서만 유일하면 된다 — 닫힌 surface 는 그 프레임에
-    /// 안 보이므로 `endFrame` 이 매핑을 놓아주고, 같은 주소가 재사용돼도 새 매핑으로 시작한다.
-    surface: usize,
+    /// surface 의 **재사용되지 않는** 식별자(`Surface.id` — `SurfaceIdAllocator` 가 단조 증가로 발급).
+    /// **포인터를 쓰면 안 된다**: surface 가 닫히고 같은 주소에 새 surface 가 두 프레임 사이에 할당되면 새
+    /// surface 가 옛 매핑(같은 전역 id)을 그대로 이어받는다. 그때 새 이미지가 `id 1·generation 1` 처럼 옛
+    /// 것과 같은 generation 이면(코어마다 1 부터라 가장 흔하다) 업로드 기록이 「이미 올렸다」고 답해 **옛
+    /// surface 의 그림이 보인다**. 비재사용 id 는 새 surface 를 늘 새 키로 만든다.
+    surface: u64,
     local: u32,
 };
